# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crypto Strategy Lab API
  version: 0.2.0
  description: |-
    Draft public REST contract for the Crypto Strategy Lab MVP.
    Long-running Backtest and Experiment operations are asynchronous.
    WebSocket messages are versioned separately in `websocket-events.md`.
servers:
  - url: /api/v1
    description: Current host
security:
  - bearerAuth: []
...
paths:
...
  /experiments:
    post:
      tags: [Experiments]
      operationId: startExperiment
      summary: Start a strategy search Experiment
      description: |-
        Reserved contract for freezing a search configuration and queuing execution.
        This operation currently returns `503 DEPENDENCY_UNAVAILABLE` until the Search
        Coordinator publishes its application boundary; no partial Experiment is created.
      x-readiness: BLOCKED_SEARCH_COORDINATOR
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartExperimentRequest' }
            example:
              name: BTC MA random search
              datasetId: 01JDATASET00000000000000001
              generator: { generatorId: random, version: 1.0.0, seed: 20260816 }
              searchSpace:
                strategyId: ma-crossover
                strategyVersion: 1.0.0
                parameters:
                  fastPeriod: { minimum: 5, maximum: 30 }
                  slowPeriod: { minimum: 40, maximum: 120 }
              stopCondition: { maximumCandidates: 100, maximumDurationSeconds: 300 }
              topK: 10
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /experiments/{experimentId}:
    get:
      tags: [Experiments]
      operationId: getExperiment
      summary: Get an authoritative Experiment snapshot
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ExperimentId'
      responses:
        '200':
          description: Current durable Experiment state.
          headers:
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Experiment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /experiments/{experimentId}/stop:
    post:
      tags: [Experiments]
      operationId: stopExperiment
      summary: Request an owned Experiment to stop
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Stop request accepted or returned from an equivalent replay.
          headers:
            Location:
              description: URL of the authoritative Experiment state.
              schema: { type: string, format: uri-reference }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Experiment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /experiments/{experimentId}/reproductions:
    post:
      tags: [Experiments]
      operationId: reproduceExperiment
      summary: Create a new Experiment linked to immutable original evidence
      description: |-
        Reserved contract. This operation currently returns `503 DEPENDENCY_UNAVAILABLE`
        until the Search Coordinator is available; the original Experiment is not modified.
      x-readiness: BLOCKED_SEARCH_COORDINATOR
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /experiments/{experimentId}/candidates:
    get:
      tags: [Experiments]
      operationId: listExperimentCandidates
      summary: List deterministic Candidate definitions for an owned Experiment
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Candidate page returned successfully.
          headers:
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidatePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /experiments/{experimentId}/candidates/{candidateId}:
    get:
      tags: [Experiments]
      operationId: getExperimentCandidate
      summary: Get one Candidate through its owned Experiment parent
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/CandidateId'
      responses:
        '200':
          description: Candidate returned successfully.
          headers:
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Candidate' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  headers:
    CorrelationId:
      description: Correlation identifier generated or accepted by the API.
      schema: { $ref: '#/components/schemas/OpaqueId' }
    RetryAfter:
      description: Seconds the client should wait before retrying.
      schema: { type: integer, minimum: 1 }
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Optional client correlation ID; an invalid value may be replaced.
      schema: { $ref: '#/components/schemas/OpaqueId' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Reusing this key with a different canonical payload returns `409 Conflict`.
      schema: { type: string, minLength: 1, maxLength: 255 }
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema: { type: string, minLength: 1 }
...
    ExperimentId:
      name: experimentId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
    CandidateId:
      name: candidateId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
  responses:
    BadRequest:
      description: Request syntax, schema, query parameter, or cursor is invalid.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication is missing or invalid; token validation details are concealed.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: A requested or referenced resource was not found.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Resource state, version, immutability, or idempotency conflict.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnsupportedMediaType:
      description: Request Content-Type is unsupported.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableContent:
      description: Structurally valid request violates a domain rule.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Client rate limit or quota exceeded.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: A provider, database, queue, or internal service is temporarily unavailable.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 128
      description: Opaque public identifier; clients must not parse it.
...
    PaginationFields:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, 'null'], minLength: 1 }
        hasMore: { type: boolean }
...
    GeneratorSelection:
      type: object
      additionalProperties: false
      required: [generatorId, version, seed]
      properties:
        generatorId: { type: string, minLength: 1 }
        version: { type: string, minLength: 1 }
        seed: { type: integer, format: int64 }
    NumericRange:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        minimum: { type: number }
        maximum: { type: number }
        step: { type: number, exclusiveMinimum: 0 }
    SearchSpace:
      type: object
      additionalProperties: false
      required: [strategyId, strategyVersion, parameters]
      properties:
        strategyId: { type: string, minLength: 1 }
        strategyVersion: { type: string, minLength: 1 }
        parameters:
          type: object
          minProperties: 1
          additionalProperties: { $ref: '#/components/schemas/NumericRange' }
    StopCondition:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        maximumCandidates: { type: integer, minimum: 1 }
        maximumDurationSeconds: { type: integer, minimum: 1 }
    StartExperimentRequest:
      type: object
      additionalProperties: false
      required: [name, datasetId, generator, searchSpace, stopCondition, topK]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        datasetId: { $ref: '#/components/schemas/OpaqueId' }
        generator: { $ref: '#/components/schemas/GeneratorSelection' }
        searchSpace: { $ref: '#/components/schemas/SearchSpace' }
        stopCondition: { $ref: '#/components/schemas/StopCondition' }
        topK: { type: integer, minimum: 1, maximum: 100 }
...
    ExperimentStatus:
      type: string
      enum: [CREATED, QUEUED, RUNNING, STOP_REQUESTED, STOPPED, COMPLETED, FAILED]
    ExperimentFailure:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        message: { type: string, minLength: 1 }
    Experiment:
      type: object
      additionalProperties: false
      required: [experimentId, name, status, datasetId, jobIds, derivedFromExperimentId, reproducesExperimentId, startedAt, completedAt, failure, createdAt]
      properties:
        experimentId: { $ref: '#/components/schemas/OpaqueId' }
        name: { type: string, minLength: 1 }
        status: { $ref: '#/components/schemas/ExperimentStatus' }
        datasetId: { $ref: '#/components/schemas/OpaqueId' }
        jobIds:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/OpaqueId' }
        derivedFromExperimentId: { type: [string, 'null'], minLength: 1 }
        reproducesExperimentId: { type: [string, 'null'], minLength: 1 }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        failure:
          oneOf:
            - $ref: '#/components/schemas/ExperimentFailure'
            - type: 'null'
        createdAt: { type: string, format: date-time }
    Candidate:
      type: object
      additionalProperties: false
      required: [candidateId, experimentId, generationIndex, definition, generatorState, fingerprint, createdAt]
      properties:
        candidateId: { $ref: '#/components/schemas/OpaqueId' }
        experimentId: { $ref: '#/components/schemas/OpaqueId' }
        generationIndex: { type: integer, minimum: 0 }
        definition: { type: object }
        generatorState: { type: [object, 'null'] }
        fingerprint: { type: string, minLength: 1 }
        createdAt: { type: string, format: date-time }
    CandidatePage:
      allOf:
        - $ref: '#/components/schemas/PaginationFields'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Candidate' }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1 }
    ErrorDetails:
      type: object
      additionalProperties: true
      properties:
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        resourceType: { type: string }
        resourceId: { $ref: '#/components/schemas/OpaqueId' }
        currentState: { type: string }
        allowedStates:
          type: array
          items: { type: string }
        retryable: { type: boolean }
        retryAfterSeconds: { type: [integer, 'null'], minimum: 1 }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, details, correlationId, timestamp]
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        message: { type: string, minLength: 1 }
        details: { $ref: '#/components/schemas/ErrorDetails' }
        correlationId: { $ref: '#/components/schemas/OpaqueId' }
        timestamp: { type: string, format: date-time }
      example:
        code: STRATEGY_PARAMETERS_INVALID
        message: Strategy parameters are invalid.
        details:
          fieldErrors:
            - field: strategy.parameters.fastPeriod
              reason: must be less than slowPeriod
          retryable: false
        correlationId: 01JBACKTESTREQUEST0000000001
        timestamp: '2026-08-16T02:05:00Z'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalServiceToken:
      type: http
      scheme: bearer
      description: Rotated internal service token; never exposed to browser clients.